requirements.txt
aws-cdk-lib==2.202.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
pandas>=1.5.0
scikit-learn>=1.1.0
numpy>=1.21.0
requests>=2.32.4